Docs: LocalTrue는 지역off,LocalFalse는 지역on으로 파일이름 변경
