+++ src/pages/Post/element/LocalFalse/HelpUsFalse.js
@@ -1,18 +1,23 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
 
 const HelpUsFalse = () => {
   const dispatch = useDispatch();
-  // const real = useSelector((state))
-
-
+  // const real = useState((state))
 
   // useEffect()
+
   return (<>
-    <StHelpeeWrapper>
-      헬퍼스(로컬)
-    </StHelpeeWrapper>
+    <StHelpeUsWrapper>
+      헬퍼스
+    </StHelpeUsWrapper>
   </>)
 }
 export default HelpUsFalse;
+
+const StHelpeUsWrapper = styled.div`
+width:95%;
+margin:auto;
+  border:1px solid #000;
+`
