+++ src/pages/Post/element/LocalFalse/HelpeeFalse.js
@@ -11,8 +11,14 @@ const HelpeeFalse = () => {
   // useEffect()
   return (<>
     <StHelpeeWrapper>
-      헬피(로컬)
+      헬피
     </StHelpeeWrapper>
   </>)
 }
 export default HelpeeFalse;
+
+const StHelpeeWrapper = styled.div`
+width:95%;
+margin:auto;
+  border:1px solid #000;
+`
