+++ src/pages/Post/element/LocalTrue/HelpeeTrue.js
@@ -2,22 +2,17 @@ import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
 
-
-const HelperFalse = () => {
+const HelpeeFalse = () => {
   const dispatch = useDispatch();
   // const real = useSelector((state))
 
+
+
   // useEffect()
   return (<>
-    <StHelperWrapper>
-      헬퍼
-    </StHelperWrapper>
+    <StHelpeeWrapper>
+      헬피(로컬)
+    </StHelpeeWrapper>
   </>)
 }
-export default HelperFalse;
-
-const StHelperWrapper = styled.div`
-width:95%;
-margin:auto;
-  border:1px solid #000;
-`
+export default HelpeeFalse;
